Refactored log_report.py into 2 scripts (kind of like MVC pattern)
