+++ utils/log_report.py
@@ -1,558 +1,129 @@
 #!/usr/bin/env python
 
+"""
+Plays the role of both "view" (report) and "controller" (main)
+"""
 
-import argparse, collections, subprocess, re, os
-from datetime import datetime, timedelta
+import argparse
+
+from create_job_records import JobLog, create_job_logs
 
 
 # All dates are printed in this format, eg. '2016-01-01T09:02:03'
-# Path to the log file for this job
 OUTPUT_DATE_FORMAT = '%Y-%m-%dT%H-%M-%S'
 
+# If a field is undefined, represent it as follows:
+UNDEFINED = 'N/A'
 
-####################################################################################################################
-# CLASSES
-####################################################################################################################
 
-# Data container for the information associated with a job
-class JobLog:
-    # These are the fields in a JobLog
-    __slots__ = ['job_id', 'job_full_id', 'job_name', 'job_dependencies',
-                 'status', 'exit_status', 'MUGQIC_exit_status',
-                 'walltime', 'start_date', 'end_date', 'cput', 'cpu_to_real_time_ratio',
-                 'mem', 'vmem', 'vmem_to_mem_ratio', 'limits', 'queue',
-                 'username', 'group', 'session', 'account',
-                 'nodes', 'path']
+###################################################################################################
+# SUMMARIZE JOBS
+###################################################################################################
 
-    # You can initialize a JobLog via keyword arguments, eg:
-    # JobLog(job_id='123', mem='40Gb', job_dependencies = ['456', '789'])
-    def __init__(self, **kwargs):
-        for key in kwargs:
-            setattr(self, key, kwargs[key])
+def min_max_field(job_logs, field):
+    """
+    Return name and field of the min and max JobLogs with respect to field
+    Return UNDEFINED * 4 if we can't find values
+    :param field: JobLog field
+    :return: min_name, min_field, max_name, max_field strings
+    """
+    # Sort the jobs by 'field'
+    jobs = sorted(filter(lambda log: hasattr(log, field), job_logs), key=lambda log: getattr(log, field))
+    if len(jobs) > 0:
+        min_job = jobs[0]
+        max_job = jobs[-1]
+        return min_job.job_name, str(getattr(min_job, field)), max_job.job_name, str(getattr(max_job, field))
+    else:
+        return UNDEFINED, UNDEFINED, UNDEFINED, UNDEFINED
 
-    def format_field(self, field):
-        """
-        Lookup the field in self, and format the result for displaying in a table
 
-        :param field: one of the fields in __slots__
-        :return: string representation of self's field
-        """
-        if hasattr(self, field):
-            if field == 'start_date' or field == 'end_date':
-                # Always format dates as OUTPUT_DATE_FORMAT
-                return str(getattr(self, field).strftime(OUTPUT_DATE_FORMAT))
-            else:
-                return str(getattr(self, field))
+def summarize_status(job_logs):
+    """
+    return the number of jobs that have status:
+             SUCCESS, ACTIVE, INACTIVE, FAILED
+    If 'status' is not defined on all jobs, return UNDEFINED's instead
+    """
+    # If every job has a status, return a breakdown of the jobs
+    if len(filter(lambda log: hasattr(log, 'status'), job_logs)) > 0:
+        return sum(log.status == 'SUCCESS' for log in job_logs), \
+               sum(log.status == 'ACTIVE' for log in job_logs), \
+               sum(log.status == 'INACTIVE' for log in job_logs), \
+               sum(log.status == 'FAILED' for log in job_logs)
+    else:
+        return UNDEFINED, UNDEFINED, UNDEFINED, UNDEFINED
+
+
+def get_start_date(job_logs):
+    start_dates = [log.start_date for log in job_logs if hasattr(log, 'start_date')]
+
+    if len(start_dates) > 0:
+        return min(start_dates)
+    return UNDEFINED
+
+
+def get_end_date(job_logs):
+    end_dates = [log.end_date for log in job_logs if hasattr(log, 'end_date')]
+    if len(end_dates) > 0:
+        return min(end_dates)
+    return UNDEFINED
+
+
+###################################################################################################
+# FORMAT LOG
+###################################################################################################
+
+def format_field(log, field):
+    """
+    Lookup the field in log, and format the result for displaying in a table
+
+    :param field: one of the fields of a JobLog
+    :return: string representation of log's field
+    """
+    if hasattr(log, field):
+        if field == 'start_date' or field == 'end_date':
+            # Always format dates as OUTPUT_DATE_FORMAT
+            return str(getattr(log, field).strftime(OUTPUT_DATE_FORMAT))
         else:
-            # If the field is not defined, display the field as 'N/A'
-            return 'N/A'
+            return str(getattr(log, field))
+    else:
+        # If the field is not defined, display the field as 'N/A'
+        return UNDEFINED
 
-    # Return a tab-delimited list in order of __slots__
+
+def format_log(log):
+    """
+    :param log: JobLog
+    :return: tab-delimited string in order of JobLog.__slots__
+    """
     # Eg:
     # 123   456 trimmomatic 789 SUCCESS 0   0   ...
-    def __str__(self):
-        return '\t'.join(self.format_field(field) for field in JobLog.__slots__)
+    return '\t'.join(format_field(log, field) for field in JobLog.__slots__)
 
 
-class JobDependencies:
-    """
-    Data container for the dependencies of a job
-    The actual values are stored as a list of job ids (eg. ['123', '456'])
-    """
-    def __init__(self, dependency_string):
-        self.dependencies = dependency_string.split(':')
-
-    def __str__(self):
-        return ':'.join(str(dep) for dep in self.dependencies)
-
-
-class RE:
-    """
-    Wrapper to make regex matching easier
-    Do the 'search', then access the results with 'group'
-    """
-    @classmethod
-    def search(cls, pattern, string, flag=0):
-        m = re.search(pattern, string, flag)
-        if m != None:
-            cls.groups = m.groups()
-            return True
-
-        return False
-
-    @classmethod
-    def group(cls, num):
-        return cls.groups[num - 1]
-
-
-class MemorySize:
-    """
-    Class to keep track of memory sizes (eg. '123kb') and be useful for displaying
-
-    Use the 'bytes' int field
-    """
-    def __init__(self, memory_size):
-        """
-        If no unit is given, bytes are assumed
-
-        :param memory_size: formatted string (eg. '2M')
-        """
-        if RE.search('^(\d*\.?\d*)g', memory_size, re.IGNORECASE):
-            self.bytes = int(float(RE.group(1)) * 1024 ** 3)
-        elif RE.search('^(\d*\.?\d*)m', memory_size, re.IGNORECASE):
-            self.bytes = int(float(RE.group(1)) * 1024 ** 2)
-        elif RE.search('^(\d*\.?\d*)k', memory_size, re.IGNORECASE):
-            self.bytes = int(float(RE.group(1)) * 1024)
-        elif RE.search('^(\d+)b', memory_size, re.IGNORECASE):
-            self.bytes = int(float(RE.group(1)))
-        else:
-            raise Exception('Can\'t convert ' + memory_size + ' to a value in bytes')
-
-    # The default representation of a memory size in Gb
-    def __str__(self):
-        return '{0:.2f}'.format(float(self.bytes) / 1024 ** 3) + ' GiB'
-
-    def __repr__(self):
-        return self.__str__()
-
-
-####################################################################################################################
-# HELPERS
-####################################################################################################################
-
-def conditional_assign(obj, attribute_name, value):
-    """
-    Assign obj.attribute_name = value if the attribute is not defined
-    """
-    if not hasattr(obj, attribute_name):
-        setattr(obj, attribute_name, value)
-
-
-def run_command(cmd_list):
-    """
-    E.g. ['showjobs', '-o', '-u', 'nreinhardt']
-    :param cmd_list: list of command, options, and values
-    :return: stdout of command results, with leading and trailing whitespace removed
-    :raises: Exception if command fails
-    """
-    process = subprocess.Popen(cmd_list, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    output, err = process.communicate()
-    ret_code = process.wait()
-
-    if ret_code != 0 or output.strip() == '':
-        raise Exception('Command ' + str(cmd_list) + ' failed or did not return output')
-
-    return output.strip()
-
-
-def get_file_date(filename):
-    """
-    :return: seconds since epoch of date file status last changed
-    """
-    return int(os.path.getctime(filename))
-
-
-def get_user():
-    try:
-        return run_command(['whoami'])
-    except:
-        return None
-
-
-# Necessary for Python versions < 2.7
-def to_seconds(td):
-    """
-    :param td: datetime.timedelta
-    :return: int
-    """
-    return td.seconds + td.days * 24 * 3600
-
-
-def percent(num):
-    """
-    Format num as a percent
-    :return: str
-    """
-    return '{0:.0f}%'.format(100 * num)
-
-
-####################################################################################################################
-# JOB LIST FILE
-####################################################################################################################
-
-def parse_job_list_file(filename):
-    """
-    :param filename: path to job list file
-    :return: list of JobLogs
-    """
-    job_logs = []
-
-    with open(filename) as f:
-        for line in f.readlines():
-            id, name, dependency_string, log_file_path = line.strip().split('\t')
-
-            dependencies = JobDependencies(dependency_string)
-
-            # Get the directory of the job list file, and then add on the relative path in the job list file
-            # to get the full path to the log file
-            full_log_file_path = os.path.abspath(os.path.join(os.path.dirname(filename), log_file_path))
-
-            job_logs.append(JobLog(job_name=name, job_id=id, job_dependencies=dependencies, path=full_log_file_path))
-
-    return job_logs
-
-
-####################################################################################################################
-# CLUSTER JOB LOG
-####################################################################################################################
-
-def parse_cluster_job_log_paths(job_output_dir, job_logs):
-    """
-    Get the MUGQICexitStatus from each job's full log
-
-    :param job_output_dir: path to directory containing all the job logs
-    :param job_logs: list of JobLogs
-    :return: modified list of JobLogs
-    """
-    for log in job_logs:
-        if os.path.exists(log.path):
-            # If the output file exists, then the job must have ended at the date the file was created
-            log.end_date = datetime.fromtimestamp(get_file_date(log.path))
-
-            with open(log.path) as f:
-                for line in f.readlines():
-                    if RE.search('MUGQICexitStatus:(\S+)', line):
-                        log.MUGQIC_exit_status = RE.group(1)
-
-    return job_logs
-
-
-####################################################################################################################
-# FILTER BY SUCCESS
-####################################################################################################################
-
-def filter_by_success(job_logs, keep_successful, keep_unsuccessful):
-    """
-    :param job_logs: list of JobLogs
-    :param keep_successful: whether to keep successful jobs
-    :param keep_unsuccessful: whether to keep jobs that have not succeeded yet (blocked or failed)
-    :return: only those JobLogs that search the given criteria
-    """
-    successful_job = lambda log: log.MUGQIC_exit_status == 0
-
-    if keep_successful:
-        job_logs = [log for log in job_logs if successful_job(log)]
-    elif keep_unsuccessful:
-        job_logs = [log for log in job_logs if not successful_job(log)]
-
-    return job_logs
-
-
-####################################################################################################################
-# ASSIGN COMPUTED FIELDS
-####################################################################################################################
-
-def assign_status(job_logs):
-    for log in job_logs:
-        if hasattr(log, 'MUGQIC_exit_status') and log.MUGQIC_exit_status == '0':
-            log.status = 'SUCCESS'
-        elif hasattr(log, 'end_date'):
-            log.status = 'FAILED'
-        elif hasattr(log, 'start_date'):
-            log.status = 'ACTIVE'
-        else:
-            log.status = 'INACTIVE'
-
-    return job_logs
-
-
-def assign_cpu_to_real_time_ratio(job_logs):
-    """
-    Assign "cput / walltime" ratio to each JobLog if possible
-    :param job_logs: list of JobLogs
-    :return: modified list of JobLogs
-    """
-    for log in job_logs:
-        if hasattr(log, 'cput') and hasattr(log, 'walltime') and log.walltime != 0:
-            log.cpu_to_real_time_ratio = percent(log.cput.total_seconds() / log.walltime.total_seconds())
-
-    return job_logs
-
-
-def assign_vmem_to_mem_ratio(job_logs):
-    """
-    Assign "vmem / mem" ratio to each JobLog if possible
-    :param job_logs: list of JobLogs
-    :return: modified list of JobLogs
-    """
-    for log in job_logs:
-        if hasattr(log, 'vmem') and hasattr(log, 'mem') and log.mem.bytes != 0:
-            log.vmem_to_mem_ratio = percent(float(log.vmem.bytes) / float(log.mem.bytes) - 1)
-
-    return job_logs
-
-
-####################################################################################################################
-# SHOWJOBS PARSING
-####################################################################################################################
-
-def get_end_date(job_log_list):
-    """
-    :param job_log_list: list of JobLogs
-    :return: the maximum end date of all jobs in format YYYY-MM-DD, or None if we can't find it
-    """
-    # The end dates of all the jobs in Unix timestamp
-    try:
-        end_date_timestamps = [int(get_file_date(log.path)) for log in job_log_list]
-    except:
-        return None
-
-    # Return the maximum date
-    end_date = max(datetime.fromtimestamp(stamp) for stamp in end_date_timestamps)
-
-    return end_date.strftime('%Y-%m-%d')
-
-
-def get_all_showjobs_output(job_list_filename, job_log_list):
-    """
-    :param job_list_filename: path to the job list file
-    :param job_log_list: list of JobLogs
-    :return: string containing all possible showjobs output for this job log
-    """
-    # Get the end date of the latest job
-    query_start_date = datetime.fromtimestamp(int(get_file_date(job_list_filename))).strftime('%Y-%m-%d')
-    # Options must be passed as a list of strings
-    start_date_option = ['-s', query_start_date] if query_start_date else []
-
-    # Get the earliest start date
-    query_end_date = get_end_date(job_log_list)
-    end_date_option = ['-e', query_end_date] if query_end_date else []
-
-    user = get_user()
-    user_option = ['-u', str(user)] if user else []
-
-    # Run a 'showjobs' query on the specified date range for this user
-    try:
-        showjobs_results = run_command(['showjobs'] + start_date_option + end_date_option + user_option)
-    except:
-        return None
-
-    return showjobs_results
-
-
-def get_showjobs_output(job_list_filename, job_logs):
-    """
-    :param job_list_filename: path to the job list file
-    :param job_logs: list of JobLogs
-    :return: dict mapping from job id -> list of lines of relevant showjobs results
-    """
-    id_to_entry = dict()
-
-    all_results = get_all_showjobs_output(job_list_filename, job_logs)
-    if all_results == None:
-        return None
-
-    ids = set([log.job_id for log in job_logs])
-
-    for entry in all_results.split('-' * 80):
-        entry_lines = entry.splitlines()
-
-        for line in entry_lines:
-            if RE.search('Job Id *: (\d+)', line): # We found the job id line
-                id = RE.group(1)
-
-                # Don't save this output entry unless this id is one of the ids we are looking for
-                if id in ids:
-                    id_to_entry[id] = entry_lines
-                break
-
-    return id_to_entry
-
-
-def parse_showjobs_output(id_to_showjobs_entry, job_log):
-    """
-    Add fields to job_log by parsing the showjobs output
-
-    :param id_to_showjobs_entry: dict mapping from job id -> list of showjobs output lines
-    :return: modified JobLog
-    """
-    # Get the relevant part of the showjobs output for this job
-    entry = id_to_showjobs_entry[job_log.job_id]
-
-    for line in entry:
-        if RE.search('^Job Id *: (\d+)$', line):
-            job_log.job_full_id = RE.group(1)
-        elif RE.search('^Start Time *: (.*)$', line):
-            job_log.start_date = datetime.strptime(RE.group(1), '%a %b %d %X %Y')
-        elif RE.search('^User Name *: (\S+)$', line):
-            job_log.username = RE.group(1)
-        elif RE.search('^Group Name *: (\S+)$', line):
-            job_log.group = RE.group(1)
-        elif RE.search('^CPUTime *: (\d+):(\d+):(\d+)$', line):
-            job_log.cput = timedelta(hours=int(RE.group(1)), minutes=int(RE.group(2)), seconds=int(RE.group(3)))
-        elif RE.search('^Memory Used *: (\S+)$', line):
-            job_log.mem = MemorySize(RE.group(1))
-        elif RE.search('^vmem Used *: (\S+)$', line):
-            job_log.vmem = MemorySize(RE.group(1))
-        elif RE.search('^Wallclock Duration *: (\d+):(\d+):(\d+)$', line):
-            job_log.walltime = timedelta(hours=int(RE.group(1)), minutes=int(RE.group(2)), seconds=int(RE.group(3)))
-        elif RE.search('^Queue Name *: (\S+)$', line):
-            job_log.queue = RE.group(1)
-        elif RE.search('^Exit Code *: (\S+)$', line):
-            job_log.exit_status = RE.group(1)
-        elif RE.search('^End Time *: (.*)$', line):
-            job_log.end_date = datetime.strptime(RE.group(1), '%a %b %d %X %Y')
-
-    return job_log
-
-
-####################################################################################################################
-# CHECKJOB PARSING
-####################################################################################################################
-
-def get_checkjob_output(job_id):
-    '''
-    :return: list of lines containing checkjob output for this job
-    '''
-    try:
-        return run_command(['checkjob', '-v', str(job_id)]).splitlines()
-    except:
-        return ''
-
-
-def parse_checkjob_output(checkjob_results, job_log):
-    """
-    Add fields to job_log by parsing the showjobs output
-    Only assign fields if they have not been defined already (don't overwrite)
-
-    :param checkjob_results: list of lines containing checkjob output for this job
-    :return: modified job_log
-    """
-    for index, line in enumerate(checkjob_results):
-        if RE.search('^job (\d+)$', line):
-            conditional_assign(job_log, 'job_full_id', RE.group(1))
-        elif RE.search('^Completion Code: (\d+) *Time: (\S+)$', line):
-            conditional_assign(job_log, 'exit_status', RE.group(1))
-            # We have to add in the current year manually
-            conditional_assign(job_log, 'end_date', datetime.strptime(RE.group(2) + str(datetime.now().year), '%a %b %d %H:%M:%S %Y'))
-        elif RE.search('^Creds: *user:(\S+) *group:(\S+) *class:(\S+)', line):
-            conditional_assign(job_log, 'username', RE.group(1))
-            conditional_assign(job_log, 'group', RE.group(2))
-        # Only assign the walltime if the job has finished
-        elif RE.search('^WallTime:\s+(\d+):(\d+):(\d+)', line) and hasattr(job_log, 'MUGQIC_exit_status'):
-            conditional_assign(job_log, 'walltime', timedelta(hours=RE.group(1), minutes=RE.group(2), seconds=RE.group(3)))
-        elif RE.search('^StartTime: *(\S+)$', line):
-            conditional_assign(job_log, 'start_date', datetime.strptime(RE.group(1) + str(datetime.now().year), '%a %b %d %H:%M:%S %Y'))
-        elif RE.search('^Allocated Nodes:', line):
-            # This field's value is on the following line, for example
-            # Allocated Nodes:
-            # [r2a - 1:3][r3a - 1:3][r4a - 1:3]
-            next_line = checkjob_results[index + 1]
-            conditional_assign(job_log, 'nodes', next_line)
-        elif RE.search('^OuptutFile: *(\S+):(//\S+)$', line):
-            conditional_assign(job_log, 'path', os.path.abspath(RE.group(1)))
-        elif RE.search('^Submit Args:', line) and not hasattr(job_log, 'limits'):
-            if RE.search('(nodes:\d+)', line):
-                job_log.limits = RE.group(1)
-            # Add more colon-delimited information to the limits field
-            if RE.search('(ppn=\d+)', line):
-                job_log.limits = ':' + job_log.limits if hasattr(job_log, 'limits') else RE.group(1)
-
-    return job_log
-
-
-####################################################################################################################
-# POPULATE JOB LOGS
-####################################################################################################################
-
-def populate_job_logs(job_list_file, keep_successful, keep_unsuccessful):
-    """
-    Aggregate information about each job and return a list of JobLogs
-
-    :param job_list_file: path to the job list file
-    :param keep_successful: whether to keep the successful jobs
-    :param keep_unsuccessful:  whether to keep the unsuccessful or blocked jobs
-    :return: list of JobLogs
-    """
-    job_logs = parse_job_list_file(job_list_file)
-
-    # Path to the directory containing all the job logs
-    job_output_dir = os.path.dirname(job_list_file)
-
-    # Go into each job's full log, and extract the results
-    job_logs = parse_cluster_job_log_paths(job_output_dir, job_logs)
-
-    # Remove the successful/unsuccessful jobs as dictated by the given options
-    job_logs = filter_by_success(job_logs, keep_successful, keep_unsuccessful)
-
-    # dict mapping from id -> list of lines of relevant showjobs results
-    id_to_showjobs_entry = get_showjobs_output(job_list_file, job_logs)
-
-    for log in job_logs:
-        if id_to_showjobs_entry:
-            log = parse_showjobs_output(id_to_showjobs_entry, log)
-
-        checkjob_results = get_checkjob_output(log.job_id)
-        log = parse_checkjob_output(checkjob_results, log)
-
-
-    job_logs = assign_status(job_logs)
-    job_logs = assign_cpu_to_real_time_ratio(job_logs)
-    job_logs = assign_vmem_to_mem_ratio(job_logs)
-
-
-    return job_logs
-
-
-####################################################################################################################
-# SUMMARIZE JOBS
-####################################################################################################################
-
-def summarize_jobs(job_logs):
-    """
-    Compute a summary of the job_logs, like start date, highest memory, etc.
-
-    :param job_logs: list of JobLogs
-    :return: start date, end date,
-             lowest memory JobLog, highest memory JobLog,
-             shortest JobLog, longest JobLog,
-             and the number of jobs that are successful, active, inactive, and failed,
-    """
-    # Sort the jobs that have a 'mem' attribute
-    sorted_by_memory = sorted(filter(lambda log: hasattr(log, 'mem'), job_logs), key = lambda log: log.mem)
-    # Sort the jobs that have a 'walltime' attribute
-    sorted_by_time = sorted(filter(lambda log: hasattr(log, 'walltime'), job_logs), key = lambda log: log.walltime)
-
-    return min([log.start_date for log in job_logs if hasattr(log, 'start_date')]), \
-           max([log.end_date for log in job_logs if hasattr(log, 'end_date')]), \
-           sorted_by_memory[0], \
-           sorted_by_memory[-1], \
-           sum(log.status == 'SUCCESS' for log in job_logs), \
-           sum(log.status == 'ACTIVE' for log in job_logs), \
-           sum(log.status == 'INACTIVE' for log in job_logs), \
-           sum(log.status == 'FAILED' for log in job_logs), \
-           sorted_by_time[0], \
-           sorted_by_time[-1]
-
-
-####################################################################################################################
+###################################################################################################
 # REPORT
-####################################################################################################################
+###################################################################################################
 
-def get_log_text_report(job_logs):
+def get_log_text_report(job_logs, minimal_detail=False):
     """
     Unassigned fields are given the value 'N/A'
     :return: report string
     """
-    start_date, end_date, lowest_memory_job, highest_memory_job,\
-        num_successful, num_active, num_inactive, num_failed,\
-        shortest_job, longest_job = summarize_jobs(job_logs)
+    report = ''
 
-    exec_time = start_date.strftime(OUTPUT_DATE_FORMAT) + ' - ' + end_date.strftime(OUTPUT_DATE_FORMAT) +\
-                ' (' + str(end_date - start_date) + ')'
+    if not minimal_detail:
+        start_date = get_start_date(job_logs)
+        end_date = get_end_date(job_logs)
 
+        min_mem_name, min_mem, max_mem_name, max_mem = min_max_field(job_logs, 'mem')
+        min_walltime_name, min_walltime, max_walltime_name, max_walltime = min_max_field(job_logs, 'walltime')
 
-    summary = '''\
+        num_successful, num_active, num_inactive, num_failed = summarize_status(job_logs)
+
+        exec_time = start_date.strftime(OUTPUT_DATE_FORMAT) + ' - ' + end_date.strftime(OUTPUT_DATE_FORMAT) +\
+                    ' (' + str(end_date - start_date) + ')'
+
+        report += '''\
 # Number of jobs: {num_jobs}
 # Number of successful jobs: {successful}
 # Number of active jobs: {active}
@@ -564,53 +135,60 @@ def get_log_text_report(job_logs):
 # Shortest job: {shortest_name} ({shortest_duration})
 # Longest job: {longest_name} ({longest_duration})
 
-# Lowest memory job: {lowest_name} ({lowest_mem})
-# Highest memory job: {highest_name} ({highest_mem})
+# Lowest memory job: {min_mem_name} ({min_mem})
+# Highest memory job: {max_mem_name} ({max_mem})
 
-'''.format(num_jobs = len(job_logs), successful = num_successful, active = num_active, inactive = num_inactive, failed = num_failed,
-           exec_time = exec_time,
-           shortest_name = shortest_job.job_name, shortest_duration = str(shortest_job.walltime),
-           longest_name = longest_job.job_name, longest_duration = str(longest_job.walltime),
-           lowest_name = lowest_memory_job.job_name, lowest_mem = lowest_memory_job.mem,
-           highest_name = highest_memory_job.job_name, highest_mem = highest_memory_job.mem)
+'''.format(num_jobs=len(job_logs), successful=num_successful, active=num_active, inactive=num_inactive, failed=num_failed,
+           exec_time=exec_time,
+           shortest_name=min_walltime_name, shortest_duration=min_walltime,
+           longest_name=max_walltime_name, longest_duration=max_walltime,
+           min_mem_name=min_mem_name, min_mem=min_mem,
+           max_mem_name=max_mem_name, max_mem=max_mem)
 
-    table_header = '''\
+    report += '''\
 #JOB_ID\tJOB_FULL_ID\tJOB_NAME\tJOB_DEPENDENCIES\tSTATUS\tJOB_EXIT_CODE\tCMD_EXIT_CODE\t\
 REAL_TIME\tSTART_DATE\tEND_DATE\tCPU_TIME\tCPU_REAL_TIME_RATIO\tPHYSICAL_MEM\tVIRTUAL_MEM\t\
 EXTRA_VIRTUAL_MEM_PCT\tLIMITS\tQUEUE\tUSERNAME\tGROUP\tSESSION\tACCOUNT\tNODES\tPATH
 '''
 
-    table_body = '\n'.join([str(log) for log in job_logs])
+    report += '\n'.join([format_log(log) for log in job_logs])
 
-    return summary + table_header + table_body
+    return report
 
 
-####################################################################################################################
+###################################################################################################
 # MAIN
-####################################################################################################################
+###################################################################################################
 
 def parse_args():
     '''
     Build the argparse.ArgumentParser and parse arguments
-    :return: args from ArgumentParser
+    :return: arguments
     '''
-    arg_parser = argparse.ArgumentParser()
-    arg_parser.add_argument('job_file', help='Path to file containing jobs and associated info')
-    arg_parser.add_argument('-s', '--success', help='Show successful jobs only', action='store_true')
-    arg_parser.add_argument('-nos', '--nosuccess', help='Show unsuccessful jobs only i.e. failed or uncompleted jobs', action='store_true')
-    # arg_parser.add_argument('-v', '--verbose', help='Display some messages about what the script is doing', action='store_true')
+    parser = argparse.ArgumentParser(description='Display information about a set of jobs')
+    parser.add_argument('job_file', help='Path to file containing jobs and associated info')
+    parser.add_argument('-r', '--report', help='Display a text report summarizing the jobs and their status', action='store_true')
 
-    args = arg_parser.parse_args()
+    group = parser.add_mutually_exclusive_group()
+    group.add_argument('-s', '--success', help='Show successful jobs only', action='store_true')
+    group.add_argument('-nos', '--nosuccess', help='Show unsuccessful jobs only i.e. failed or uncompleted jobs', action='store_true')
+    group.add_argument('-m', '--mimimal_detail', help="Only aggregate minimal information about each job", action='store_true')
 
-    return args.job_file, args.success, args.nosuccess
+    parser.add_argument('-t', '--top_to_bottom', help="Display an indented text showing job dependencies from first completed to last completed", action='store_true')
+    parser.add_argument('-b', '--bottom_to_top', help="Display an indented text showing job dependencies from last completed to first completed", action='store_true')
+
+    args = parser.parse_args()
+
+    return args.job_file, args.report, args.success, args.nosuccess, args.mimimal_detail
 
 
 def main():
-    job_list_file, success_option, no_success_option = parse_args()
+    job_list_file, report, success_option, no_success_option, minimal_detail = parse_args()
 
-    job_logs = populate_job_logs(job_list_file, success_option, no_success_option)
+    job_logs = create_job_logs(job_list_file,  success_option, no_success_option, minimal_detail=minimal_detail)
 
-    print(get_log_text_report(job_logs))
+    if report:
+        print(get_log_text_report(job_logs, minimal_detail=minimal_detail))
 
 if __name__ == '__main__':
     main()
